Move ref tests into its own refs.js file
Also renamed nodeType to node within these tests
